move snow_partition_iterator and window
