+++ src/snowflake/snowpark/modin/plugin/_internal/telemetry.py
@@ -544,8 +544,8 @@ class TelemetryMeta(type):
     ) -> Union[
         "snowflake.snowpark.modin.plugin.extensions.groupby_overrides.DataFrameGroupBy",
         "snowflake.snowpark.modin.pandas.resample.Resampler",
-        "snowflake.snowpark.modin.pandas.window.Window",
-        "snowflake.snowpark.modin.pandas.window.Rolling",
+        "snowflake.snowpark.modin.plugin.extensions.window_overrides.Window",
+        "snowflake.snowpark.modin.plugin.extensions.window_overrides.Rolling",
     ]:
         """
         Metaclass for enabling telemetry data collection on class/instance methods of
@@ -557,8 +557,8 @@ class TelemetryMeta(type):
         TelemetryMeta is only set as the metaclass of:
          snowflake.snowpark.modin.plugin.extensions.groupby_overrides.DataFrameGroupBy,
          snowflake.snowpark.modin.pandas.resample.Resampler,
-         snowflake.snowpark.modin.pandas.window.Window,
-         snowflake.snowpark.modin.pandas.window.Rolling, and their subclasses.
+         snowflake.snowpark.modin.plugin.extensions.window_overrides.Window,
+         snowflake.snowpark.modin.plugin.extensions.window_overrides.Rolling, and their subclasses.
 
 
         Args:
@@ -569,8 +569,8 @@ class TelemetryMeta(type):
         Returns:
             Union[snowflake.snowpark.modin.plugin.extensions.groupby_overrides.DataFrameGroupBy,
                 snowflake.snowpark.modin.pandas.resample.Resampler,
-                snowflake.snowpark.modin.pandas.window.Window,
-                snowflake.snowpark.modin.pandas.window.Rolling]:
+                snowflake.snowpark.modin.plugin.extensions.window_overrides.Window,
+                snowflake.snowpark.modin.plugin.extensions.window_overrides.Rolling]:
                 The modified class with decorated methods.
         """
         for attr_name, attr_value in attrs.items():
